feat: Add billing address with overridden attributes
This commit introduces a `billingAddress` field to the `User` entity,
demonstrating how to override attributes of an embedded object using
`@AttributeOverrides`. The billing address fields are mapped to
different column names in the database (BILLING_STREET,
BILLING_ZIPCODE, BILLING_CITY). A new test case is added to verify
that the attribute overrides are working as expected and that both
home and billing addresses are persisted correctly.